third_party/WebKit/LayoutTests/fast/dom/shadow/shadowdom-for-media-expected.txt

CONSOLE WARNING: <shadow> doesn't work for VIDEO element host.
CONSOLE WARNING: <shadow> doesn't work for VIDEO element host.
 
PASS shadowRootForVideoWithControls is youngerShadowRootForVideoWithControls
PASS shadowRootForVideoWithControls is not oldestShadowRootForVideoWithControls
PASS shadowRootForVideoWithoutControls is youngerShadowRootForVideoWithoutControls
PASS shadowRootForVideoWithoutControls is not oldestShadowRootForVideoWithoutControls
PASS successfullyParsed is true

TEST COMPLETE
